requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
Flask==0.10.1
Flask-Slack==0.1.5
funcsigs==0.4
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
magicmock==0.3
MarkupSafe==0.23
mock==1.3.0
pbr==1.8.1
python-parsely==git+git@github.com:hbbtstar/python-parsely.git#egg=python-parsely
pytz==2015.7
requests==2.7.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
tzlocal==1.2
Werkzeug==0.11.3
wheel==0.24.0
